[UPDATE] version

+++ build.sh
@@ -1,3 +1,10 @@
+pytest tests.py
+read -p "press q to exit, ENTER to continue: " r
+if [ $r == "q" ] 
+then
+    exit 0
+fi
+    
 rm -r dist
 python setup.py sdist
 twine upload dist/*
